Ony disallow 'safe' expressions from being compared with NULL.
Though undocumented, it appears that MOD_SAFE is meant to imply that a
pointer with that attribute can never be NULL and so is always safe to
be dereferenced.
The places where it is used are:
Testing a safe value
   if (foo), foo ?x:y,  if (y || foo)... , etc
is an error, and testing the negation:
   if (!foo) ...
is an error.
Also there is a current check that testing it against another value
   if (foo == y) ...
is an error.  That doesn't seem to make sense, while testing against NULL:
   if (foo == NULL) ...
does.
So change the comparison test to only consider comparisons with NULL to be an error.

Signed-off-by: NeilBrown <neil@brown.name>