Ony disallow 'safe' expressions from being compared with NULL.
MOD_SAFE (__attribute__((safe))) marks as value a being safe to
dereference. i.e. a "non-null/non-trapping pointer"

The places where it is used are:
Testing a safe value
   if (foo), foo ?x:y,  if (y || foo)... , etc
is an error, and testing the negation:
   if (!foo) ...
is an error.

Also there is a currently check that testing it against another value
   if (foo == y) ...
is an error.  That doesn't seem justifed.  However testing against NULL:
   if (foo == NULL) ...
does.
So change the comparison test to only consider comparisons with NULL to be an error.

Signed-off-by: NeilBrown <neil@brown.name>