Allow assignment between MOD_SAFE and non-MOD_SAFE pointers.
It is perfectly valid to assign a MOD_SAFE pointer to a variable
that is not safe.

It is slightly less obvious that it is safe to assign a non-MOD_SAFE
pointer to a MOD_SAFE variable, but as

	if (unsafe)
		safe = unsafe;

will preserve the property that 'safe' is safe.  The requirements can
only be properly tested with data-flow analysis.  So rather than
report errors for correct programs, allow possibly-incorrect programs
to pass.

Signed-off-by: NeilBrown <neil@brown.name>